hey Tal, replacing window.location.href with _____ in my post is not fun; I'm going to quit R-bloggers if I find my content is modified again
view-source:https://www.r-bloggers.com/how-i-find-manage-and-use-gifs/ shows that my script was modified to

<script type="text/rocketscript">_____='https://yihui.name/en/2017/05/gifs/';</script>

I'm doing the redirect because (1) the R-bloggers pages are a horrible clutter and I hope my readers could read on totally clean pages; (2) the link back to the original post is not visually prominent; (3) I'm not comfortable with someone making money from the free content produced by the community. So please do not try to lock my posts in R-bloggers.