faq: how to completely remove official TeX Live
> This list of FAQs explain the technical details of TinyTeX for those who are curious, and alternative ways to install and use TinyTeX.

The code was tested on Ubuntu. The committed FAQ might be off-topic, but it solves the TeX Live users' pain.

> If it were not for testing TinyTeX, he would never have discovered that he had got 15GB of TeX Live in several folders (like /usr/local/texlive/2015, .../2016, and .../2017, etc.).

For Windows users, there is an uninstaller in `C:\texlive\2021\tlpkg\installer\uninst.bat` according to [this guide](http://tug.ctan.org/info/install-latex-guide-zh-cn/install-latex-guide-zh-cn.pdf), but I do not test it.